ejercicio4y5

+++ Ejercicios/4/Ejercicio4.py
@@ -1,7 +1,7 @@
 # Programa para saber si un número es par o impar
 print("Inserte su número entero: ")
 numero = input()
-comparador = float(numero)%2
+comparador = int(numero)%2
 if(comparador == 1):
     print("El número: " , numero, " es impar.")
 else:
